% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bregpcdenet_Rcpp}
\alias{bregpcdenet_Rcpp}
\title{Linear constrained Penalized regression with elastic net penalty}
\usage{
bregpcdenet_Rcpp(XY, XX, A, B, Lambda1, Lambda2, Mu, Nu, Beta0, control)
}
\arguments{
\item{XY}{X'Y where X is covariance matrix and Y response vector}

\item{XX}{X'X where X is covariance matrix}

\item{A}{linear constraint parameter A}

\item{B}{linear constraint parameter}

\item{Lambda1}{lasso penalty parameter}

\item{Lambda2}{elastic net penalty parameter}

\item{Mu}{linear constraint strength}

\item{Nu}{multiplicative factor for linear constraint strength}

\item{Beta0}{initial value of regression coefficient vector}

\item{control}{a list of parameters controling the fitting process}
}
\value{
estimated regression coefficient vector
}
\description{
Linear constrained Penalized regression with elastic net penalty
}

