% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCycleX.R
\docType{data}
\name{CellCycleX}
\alias{CellCycleX}
\title{Cell cycle data}
\format{A data frame}
\usage{
data(CellCycleX)
}
\description{
The chromatin immunoprecipitation (ChIP) data (X) contain complete binding 
information of a subset of 1790 genes for a total of 113 transcription 
factors.
}
\examples{
# data(CellCycleY) # Y
# data(CellCycleX) # X
# n <- nrow(CellCycleY); p <- ncol(CellCycleX); q <-  ncol(CellCycleY)
# control <- secure.control(spU=100/p,spV=1)
# fit.cycle <- secure.path(CellCycleY, CellCycleX, nrank = 10, nlambda = 100,
#                   orthXU = FALSE, orthV = FALSE, control = control)
}
\keyword{datasets}

