% Generated by roxygen2 (4.0.2): do not edit by hand
\name{encrypt}
\alias{decrypt}
\alias{encrypt}
\title{Encrypt and decrypt data.}
\usage{
encrypt(.name, ..., .vault = NULL)

decrypt(name, vault = NULL)
}
\arguments{
\item{.name,name}{Name of storage locker.}

\item{...}{Name-value pairs of objects to store.}

\item{.vault,vault}{Name of secure vault. If \code{NULL} looks for
\code{vault} or \code{inst/vault} in the current directory. If a string,
looks for a secure vault in the package with that name}
}
\description{
Encrypt and decrypt data.
}
\examples{
\dontrun{
encrypt("test", x = 1, y = 2)
# Encrypting to same file adds fields
encrypt("test", z = 3)

decrypt("test")
}
}

