% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{has_key}
\alias{has_key}
\title{Can you unlock the secure storage?}
\usage{
has_key(vault = NULL)
}
\arguments{
\item{vault}{Name of secure vault. If \code{NULL} looks for
\code{vault} or \code{inst/vault} in the current directory. If a string,
looks for a secure vault in the package with that name}
}
\value{
A boolean flag.
}
\description{
This ensures that we can find your private key, and you can decrypt
the encrypted master key.
}
