% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keys}
\alias{github_key}
\alias{keys}
\alias{travis_key}
\title{Retrieve public keys.}
\usage{
travis_key(repo)

github_key(username, i = 1)
}
\arguments{
\item{repo}{Travis repository name (x/y)}

\item{username}{Github username}

\item{i}{Key to use, if more than one}
}
\description{
Retrieve public keys.
}
\examples{
travis_key("hadley/dplyr")
github_key("hadley")
}

