% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_pubkey}
\alias{parse_pubkey}
\alias{parse_pubkey_string}
\title{Parse public key}
\usage{
parse_pubkey(path, type = NULL)

parse_pubkey_string(key, type = c("ssh", "ssl"))
}
\arguments{
\item{path}{Path to load public key file from}

\item{type}{Type of public key ("ssh" or "ssl"). If omitted from
\code{\link{parse_pubkey}}, will be guessed from the extension.}

\item{key}{Key as a string}
}
\description{
Parse public key
}
\examples{
\donttest{
parse_pubkey("~/.ssh/id_rsa.pub")
parse_pubkey("~/.ssh/id_rsa.pem")
}
}

