% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{secure.init}
\alias{secure.init}
\title{Initialization (internal)}
\usage{
secure.init(Y, X, U0 = NULL, V0 = NULL, D0 = NULL, nrank = 4, ort)
}
\arguments{
\item{Y}{response matrix}

\item{X}{covariate matrix}

\item{U0}{user-supplied initial value of U}

\item{V0}{user-supplied initial value of V}

\item{D0}{user-supplied initial value of D}

\item{nrank}{an integer specifying the desired rank}

\item{ort}{if TRUE, X is treated as orthogonal}
}
\value{
\item{C_ls}{least square estimate}
  \item{C_rr}{reduced rank estimate}
  \item{U0}{initial value of U}
  \item{D0}{Initial value of D}
  \item{V0}{Initial value of V}
}
\description{
initialization for secure, when no missingness in Y
}

