% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure.R
\name{secure.miss.init}
\alias{secure.miss.init}
\title{Initialization (internal)}
\usage{
secure.miss.init(Y, X, U0 = NULL, V0 = NULL, D0 = NULL, nrank,
  control = list(), method = 1)
}
\arguments{
\item{Y}{response matrix}

\item{X}{covariate matrix}

\item{U0}{user-supplied initial value of U. Currently not used.}

\item{V0}{user-supplied initial value of V. Currently not used.}

\item{D0}{user-supplied initial value of D. Currently not used.}

\item{nrank}{an integer specifying the desired rank}

\item{control}{internal control parameters}

\item{method}{method = 1, least squares; method = 2, reduced rank regression}
}
\value{
\item{C_ls}{least square estimate}
  \item{C_rr}{reduced rank estimate}
  \item{U0}{initial value of U}
  \item{D0}{Initial value of D}
  \item{V0}{Initial value of V}
}
\description{
initialization for secure when Y contains missing value
}

