% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sefar_SURR_Rcpp_ortho}
\alias{sefar_SURR_Rcpp_ortho}
\title{Sparse unit-rank regression when X is orthogonal, Rcpp version}
\usage{
sefar_SURR_Rcpp_ortho(X, Y, Au, Av, bu, bv, uk, vk, dk, nlambda, control)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{Au}{responsible for orthogonality constraint on XU}

\item{Av}{responsible for orthogonality constraint on V}

\item{bu}{responsible for orthogonality constraint on XU}

\item{bv}{responsible for orthogonality constraint on V}

\item{uk}{initial value of kth column of U}

\item{vk}{initial value of kth column of V}

\item{dk}{initial value of kth diagonal value of D}

\item{nlambda}{number of lambda to be specified in solution path}

\item{control}{a list of parameters controlling the fitting process}
}
\value{
S3 \code{sefar_SURR_Rcpp_ortho} object .....
}
\description{
Sparse unit-rank regression when X is orthogonal, Rcpp version
}

