% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skip_when_missing_key}
\alias{skip_when_missing_key}
\title{Skip tests when you can't unlock}
\usage{
skip_when_missing_key(vault = NULL)
}
\arguments{
\item{vault}{Name of secure vault. If \code{NULL} looks for
\code{vault} or \code{inst/vault} in the current directory. If a string,
looks for a secure vault in the package with that name}
}
\description{
This is useful to place at the top of tests that rely on access to secured
assets. Skipped tests do not generate an error in R CMD check etc, but
will print a visible notification.
}

