% Generated by roxygen2 (4.0.2): do not edit by hand
\name{use_secure}
\alias{use_secure}
\title{Set up a package with a secure enclave.}
\usage{
use_secure(pkg = ".")
}
\arguments{
\item{pkg}{Path to package. Defaults to working directory.}
}
\description{
This creates a \code{secure/} directory, adds it to \code{.Rbuildignore},
and adds secure to \code{Suggests}.
}

