% Generated by roxygen2 (4.0.2): do not edit by hand
\name{use_securrr}
\alias{use_securrr}
\title{Set up a package with a secure hoard}
\usage{
use_securrr(pkg = ".")
}
\arguments{
\item{pkg}{Path to package. Defaults to working directory.}
}
\description{
This creates a \code{secure/} directory and adds it to \code{.Rbuildignore}
}

