% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_user}
\alias{add_user}
\alias{remove_user}
\title{Add and remove users user.}
\usage{
add_user(name, public_key, vault = NULL)

remove_user(name, vault = ".")
}
\arguments{
\item{name}{Name of user. Currently only used to help you remember who
owns the public key}

\item{public_key}{Users public key.}

\item{vault}{Name of secure vault. If \code{NULL} looks for
\code{vault} or \code{inst/vault} in the current directory. If a string,
looks for a secure vault in the package with that name}
}
\description{
Adding or removing users will re-generate the master key and re-encrypt all
secured files.
}
\examples{
\dontrun{
# Add a github user:
add_user("hadley", github_key("hadley"))
remove_user("hadley")

# Add yourself:
add_user("hadley", local_key())
remove_user("hadley")

# Add travis user
add_user("travis", travis_key("hadley/secure"))
remove_user("travis")
}
}

