% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_user}
\alias{add_user}
\alias{remove_user}
\title{Add and remove users user.}
\usage{
add_user(name, public_key, pkg = ".")

remove_user(name, pkg = ".")
}
\arguments{
\item{name}{Name of user. Currently only used to help you remember who
owns the public key}

\item{public_key}{Users public key.}

\item{pkg}{Path to package. Defaults to working directory}
}
\description{
Adding or removing users will re-generate the master key and re-encrypt all
secured files.
}
\examples{
\dontrun{
# Add a github user:
add_user("hadley", github_key("hadley"))
remove_user("hadley")

# Add yourself:
add_user("hadley", local_key())
remove_user("hadley")

# Add travis user
add_user("travis", travis_key("hadley/secure"))
remove_user("travis")
}
}

