% Generated by roxygen2 (4.0.2): do not edit by hand
\name{encrypt}
\alias{decrypt}
\alias{encrypt}
\title{Encrypt and decrypt data.}
\usage{
encrypt(.name, ..., .pkg = ".")

decrypt(name, pkg = ".")
}
\arguments{
\item{.name,name}{Name of storage locker.}

\item{...}{Name-value pairs of objects to store.}

\item{.pkg,pkg}{Path to package. Defaults to current directory.}
}
\description{
Encrypt and decrypt data.
}
\examples{
\dontrun{
encrypt("test", x = 1, y = 2)
# Encrypting to same file adds fields
encrypt("test", z = 3)

decrypt("test")
}
}

