% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keys}
\alias{github_key}
\alias{keys}
\alias{local_key}
\alias{travis_key}
\title{Retrieve public keys.}
\usage{
travis_key(repo)

github_key(username, i = 1)

local_key(name = "id_rsa")
}
\arguments{
\item{repo}{Travis repository name (x/y)}

\item{username}{Github username}

\item{i}{Key to use, if more than one}

\item{name}{Name of key. If missing, uses first file in directory,
otherwise uses first file that matches \code{name}.}
}
\description{
Retrieve public keys.
}
\examples{
travis_key("hadley/dplyr")
github_key("hadley")
\donttest{
local_key()
}
}

