% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_secure}
\alias{use_secure}
\title{Set up a package with a secure enclave.}
\usage{
use_secure(pkg = ".")
}
\arguments{
\item{pkg}{Path to package. Defaults to working directory.}
}
\description{
This creates a \code{secure/} directory, adds it to \code{.Rbuildignore},
and adds secure to \code{Suggests}. Use \code{\link{add_user}()} to add a
new user, then \code{\link{encrypt}()} to encrypt data.
}
